finalproductmodel

+++ 27-FrontToBackSqlConnection/Areas/AdminPanel/Controllers/ProductController.cs
@@ -14,25 +14,25 @@ namespace FrontToBackSqlConnection.Areas.AdminPanel.Controllers
         private readonly AppDbContext _context;
         private readonly IWebHostEnvironment _env;
 
-        public ProductController(AppDbContext context,IWebHostEnvironment env)
+        public ProductController(AppDbContext context, IWebHostEnvironment env)
         {
             _context = context;
             _env = env;
         }
         public async Task<IActionResult> Index()
         {
-            var products=await _context.Products
-                .Where(p=>!p.isDeleted)
-                .Include(p=>p.ProductImages)
-                .Include(p=>p.Category)
-                .Select(p=>new ProductGetVM
+            var products = await _context.Products
+                .Where(p => !p.isDeleted)
+                .Include(p => p.ProductImages)
+                .Include(p => p.Category)
+                .Select(p => new ProductGetVM
                 {
-                    Id=p.Id,
-                    Name=p.Name,
-                    Price=p.Price,
-                    SKU=p.SKU,
-                    Image=p.ProductImages.FirstOrDefault().Image,
-                    CategoryName=p.Category.Name
+                    Id = p.Id,
+                    Name = p.Name,
+                    Price = p.Price,
+                    SKU = p.SKU,
+                    Image = p.ProductImages.FirstOrDefault(p => p.IsPrimary==true).Image,
+                    CategoryName = p.Category.Name
                 })
                 .ToListAsync();
 
@@ -78,13 +78,13 @@ namespace FrontToBackSqlConnection.Areas.AdminPanel.Controllers
                 return View(productCreateVM);
             }
 
-            bool exsitCategory=_context.Categories.Any(c=>c.Id==productCreateVM.CategoryId);
+            bool exsitCategory = _context.Categories.Any(c => c.Id == productCreateVM.CategoryId);
             if (!exsitCategory)
             {
                 ModelState.AddModelError(nameof(productCreateVM.CategoryId), "Category does not exist");
                 return View(productCreateVM);
             }
-            if(productCreateVM.TagIds is not null)
+            if (productCreateVM.TagIds is not null)
             {
                 bool existTag = productCreateVM.TagIds.Any(tagId => !productCreateVM.Tags.Exists(t => t.Id == tagId));
                 if (existTag)
@@ -112,16 +112,16 @@ namespace FrontToBackSqlConnection.Areas.AdminPanel.Controllers
                 Description = productCreateVM.Description,
                 SKU = productCreateVM.SKU,
                 CategoryId = productCreateVM.CategoryId.Value,
-                ProductImages=new List<ProductImage> { mainImage, hoverImage }
+                ProductImages = new List<ProductImage> { mainImage, hoverImage }
             };
 
-            if(productCreateVM.TagIds is not null)
+            if (productCreateVM.TagIds is not null)
             {
                 product.ProductTags = productCreateVM.TagIds.Select(tagId => new ProductTag { TagId = tagId }).ToList();
             }
-            string info=string.Empty;
+            string info = string.Empty;
 
-            if(productCreateVM.AdditionalPhotos is not null)
+            if (productCreateVM.AdditionalPhotos is not null)
             {
                 foreach (var file in productCreateVM.AdditionalPhotos)
                 {
@@ -140,7 +140,7 @@ namespace FrontToBackSqlConnection.Areas.AdminPanel.Controllers
                     {
                         Image = await file.CreateFile(_env.WebRootPath, "assets", "images", "website-images"),
                         IsPrimary = false
-                    }); 
+                    });
                 }
             }
             TempData["FileInfo"] = info;
@@ -152,13 +152,13 @@ namespace FrontToBackSqlConnection.Areas.AdminPanel.Controllers
             return RedirectToAction(nameof(Index));
         }
 
-        public async Task<IActionResult>Update(int? id)
+        public async Task<IActionResult> Update(int? id)
         {
-            if(id is null || id<1) return BadRequest();
+            if (id is null || id < 1) return BadRequest();
 
             Product? product = await _context.Products
-                .Include(p=>p.ProductImages)
-                .Include(p=>p.ProductTags)
+                .Include(p => p.ProductImages)
+                .Include(p => p.ProductTags)
                 .FirstOrDefaultAsync(p => p.Id == id);
             if (product is null) return NotFound();
 
@@ -171,27 +171,62 @@ namespace FrontToBackSqlConnection.Areas.AdminPanel.Controllers
                 CategoryId = product.CategoryId,
                 TagIds = product.ProductTags.Select(pt => pt.TagId).ToList(),
                 Categories = await _context.Categories.Where(c => !c.isDeleted).ToListAsync(),
-                Tags=await _context.Tags.ToListAsync(),
-                ProductImages=product.ProductImages
+                Tags = await _context.Tags.ToListAsync(),
+                ProductImages = product.ProductImages
 
             };
             return View(productUpdateVM);
-            
+
 
         }
 
         [HttpPost]
-        public async Task<IActionResult> Update(int? id,ProductUpdateVM productUpdateVM)
+        public async Task<IActionResult> Update(int? id, ProductUpdateVM productUpdateVM)
         {
-            productUpdateVM.Categories=await _context.Categories.Where(c => !c.isDeleted).ToListAsync();
-            productUpdateVM.Tags = await _context.Tags.Where(t => !t.isDeleted).ToListAsync();
+            
             if (id is null || id < 1) return BadRequest();
-            if (!ModelState.IsValid) return View(productUpdateVM);
+
 
             Product? product = await _context.Products
-                .Include(p=>p.ProductTags)
+                .Include(p => p.ProductImages)
+                .Include(p => p.ProductTags)
                 .FirstOrDefaultAsync(p => p.Id == id);
             if (product is null) return NotFound();
+
+            productUpdateVM.Categories = await _context.Categories.Where(c => !c.isDeleted).ToListAsync();
+            productUpdateVM.Tags = await _context.Tags.Where(t => !t.isDeleted).ToListAsync();
+
+            if (!ModelState.IsValid) return View(productUpdateVM);
+
+
+            if (productUpdateVM.MainPhoto is not null)
+            {
+                if (!productUpdateVM.MainPhoto.CheckFileType("image/"))
+                {
+                    ModelState.AddModelError(nameof(productUpdateVM.MainPhoto), "File type must be image");
+                    return View(productUpdateVM);
+                }
+                if (!productUpdateVM.MainPhoto.CheckFileSize(FileSize.MB, 2))
+                {
+                    ModelState.AddModelError(nameof(productUpdateVM.MainPhoto), "File size must be less than 2MB");
+                    return View(productUpdateVM);
+                }
+
+            }
+            if (productUpdateVM.HoverPhoto is not null)
+            {
+                if (!productUpdateVM.HoverPhoto.CheckFileType("image/"))
+                {
+                    ModelState.AddModelError(nameof(productUpdateVM.HoverPhoto), "File type must be image");
+                    return View(productUpdateVM);
+                }
+                if (!productUpdateVM.HoverPhoto.CheckFileSize(FileSize.MB, 2))
+                {
+                    ModelState.AddModelError(nameof(productUpdateVM.HoverPhoto), "File size must be less than 2MB");
+                    return View(productUpdateVM);
+                }
+
+            }
 
             bool existCategory = _context.Categories.Any(c => c.Id == productUpdateVM.CategoryId);
             if (!existCategory)
@@ -208,21 +243,89 @@ namespace FrontToBackSqlConnection.Areas.AdminPanel.Controllers
                     return View(productUpdateVM);
                 }
             }
-            if(productUpdateVM.Tags is null)
+            if (productUpdateVM.Tags is null)
             {
                 productUpdateVM.TagIds = new();
             }
-            _context.ProductTags.RemoveRange(product.ProductTags.Where(pTag => !productUpdateVM.TagIds.Exists(tagId => tagId == pTag.TagId)).ToList());
-            _context.ProductTags.AddRange(productUpdateVM.TagIds.Where(tId=>!product.ProductTags.Exists(pTag=>pTag.TagId==tId))
-                .ToList()
-                .Select(tId=>new ProductTag { TagId=tId,ProductId=product.Id }));
+            if (productUpdateVM.TagIds is not null)
+            {
+                _context.ProductTags.RemoveRange(product.ProductTags.Where(pTag => !productUpdateVM.TagIds.Exists(tagId => tagId == pTag.TagId)).ToList());
+                _context.ProductTags.AddRange(productUpdateVM.TagIds.Where(tId => !product.ProductTags.Exists(pTag => pTag.TagId == tId))
+                    .ToList()
+                    .Select(tId => new ProductTag { TagId = tId, ProductId = product.Id }));
+            }
+
+            if (productUpdateVM.MainPhoto is not null)
+            {
+                string fileName = await productUpdateVM.MainPhoto.CreateFile(_env.WebRootPath, "assets", "images", "website-images");
+                ProductImage mainImage = product.ProductImages.FirstOrDefault(pi => pi.IsPrimary == true);
+                mainImage.Image.DeleteFile(_env.WebRootPath, "assets", "images", "website-images");
+                product.ProductImages.Remove(mainImage);
+                product.ProductImages.Add(new ProductImage
+                {
+                    Image = fileName,
+                    IsPrimary = true
+                });
+            }
+            if (productUpdateVM.HoverPhoto is not null)
+            {
+                string fileName = await productUpdateVM.HoverPhoto.CreateFile(_env.WebRootPath, "assets", "images", "website-images");
+                ProductImage hoverImage = product.ProductImages.FirstOrDefault(pi => pi.IsPrimary == false);
+                hoverImage.Image.DeleteFile(_env.WebRootPath, "assets", "images", "website-images");
+                product.ProductImages.Remove(hoverImage);
+                product.ProductImages.Add(new ProductImage
+                {
+                    Image = fileName,
+                    IsPrimary = false
+                });
+            }
+
+            if(productUpdateVM.ImageIds is  null)
+            {
+                productUpdateVM.ImageIds = new List<int>();
+            }
+
+            var deleteImages=product.ProductImages.Where(pi=>!productUpdateVM.ImageIds.Exists(imgId=>imgId==pi.Id ) && pi.IsPrimary == false).ToList();
+            deleteImages.ForEach(di =>
+            {
+                di.Image.DeleteFile(_env.WebRootPath, "assets", "images", "website-images");
+            });
+            _context.ProductImages.RemoveRange(deleteImages);
+
+
+
+            if (productUpdateVM.AdditionalPhotos is not null)
+            {
+            string info = string.Empty;
+                foreach (var file in productUpdateVM.AdditionalPhotos)
+                {
+                    if (!file.CheckFileType("image/"))
+                    {
+                        info += $"<p class=\"text-danger\">{file.FileName} type was not correct</p>";
+                        continue;
+                    }
+
+                    if (!file.CheckFileSize(FileSize.MB, 2))
+                    {
+                        info += $"<p class=\"text-danger\">{file.FileName} size was not correct</p>";
+                        continue;
+                    }
+                    product.ProductImages.Add(new ProductImage
+                    {
+                        Image = await file.CreateFile(_env.WebRootPath, "assets", "images", "website-images"),
+                        IsPrimary = false
+                    });
+                }
+            TempData["FileInfo"] = info;
+            }
+
 
             product.Name = productUpdateVM.Name;
             product.Description = productUpdateVM.Description;
             product.Price = productUpdateVM.Price;
             product.SKU = productUpdateVM.SKU;
             product.CategoryId = productUpdateVM.CategoryId.Value;
-           
+
 
             await _context.SaveChangesAsync();
